project/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.7.0
bleach==6.0.0
calplot==0.1.7.5
category-encoders==2.6.2
CD==1.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
clearml==1.15.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
cmaes==0.9.1
cmake==3.28.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
configobj==5.0.8
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.1
distlib==0.3.8
distro==1.8.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.15.3
dvc-data==2.14.0
dvc-http==2.30.2
dvc-objects==1.0.0
dvc-render==0.5.3
dvc-studio-client==0.13.0
dvc-task==0.3.0
et-xmlfile==1.1.0
executing==1.2.0
fast-ml==3.68
fastjsonschema==2.17.1
feature-engine==1.6.2
filelock==3.12.2
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.32
grandalf==0.8
greenlet==2.0.2
huggingface-hub==0.16.4
hydra-core==1.3.2
identify==2.5.35
idna==3.4
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
iterative-telemetry==0.0.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
kombu==5.3.1
lightgbm==3.3.0
lit==17.0.6
llvmlite==0.40.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
-e git+https://gl-1.2pp.dev/data-science/credit-line.git@7a6c399e81baeed2c6cb1dbacfcbc5d3e770bcba#egg=mypackage
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openpyxl==3.1.2
optuna==3.1.0
orderedmultidict==1.0.1
orjson==3.9.5
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.2.0
plotly==5.15.0
pluggy==1.4.0
pre-commit==3.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
pygit2==1.12.2
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.8.0
pymssql==2.2.7
pyparsing==3.0.9
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.8.10
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.3.5
scikit-learn==1.3.0
scipy==1.11.1
scmrepo==1.2.1
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.42.0
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
sqltrie==0.7.0
stack-data==0.6.2
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomlkit==0.12.1
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
vine==5.0.0
virtualenv==20.25.1
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
xmltodict==0.13.0
yarl==1.9.2
zc.lockfile==3.0.post1
